pyproject.toml
[project]
name = "pinterest"
version = "0.1.0"
description = "Pet project based on pinterest"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "pydantic[email]>=2.12.5",
    "ruff>=0.15.6",
    "aiosqlite>=0.22.1",
]
